feat(ai-knowledge-base): add application count and document count to knowledge base page
- Add AiKnowledgeBaseVo class to represent knowledge base with additional counts
- Update AiKnowledgeBaseController to use AiKnowledgeBaseVo in getPage response
- Modify AiKnowledgeBaseRepository to calculate app count and document count
- Update AiKnowledgeBaseService and implementation to use new VO class
- Remove unused imports and adjust existing code to accommodate new features

+++ src/main/java/org/hzai/ai/aiapplication/repository/AiApplicationRepository.java
@@ -45,7 +45,7 @@ public class AiApplicationRepository implements PanacheRepository<AiApplication>
                 pageRequest.getSize());
     }
 
-    public Map<Long, Map<String, Object>> getAppCountByKbIdList(List<Integer> kbIdList) {
+    public Map<Long, Map<String, Object>> getAppCountByKbIdList(List<Long> kbIdList) {
         if (kbIdList == null || kbIdList.isEmpty()) {
             return Collections.emptyMap();
         }
